HWT: PWR: IO: Prepare HWT_IO_CTL to work with other device powered by FT232RL
When other FT232RL powered devices are plugged to USB, the script before
this change was not able to get access to hwt_io_ctl device.

Now it checks for multiple USB device instances with the same idProduct and
idVendor and then compare iSerialNumber, which should be HWT000{1, 2, ...}.\

The BitBangDevice() method accepts USB device name as a parameter.

Signed-off-by: Lukasz Majewski <l.majewski@samsung.com>